+++ packages/webcomponents/src/styles/parts.ts
@@ -135,6 +135,9 @@ export const paymentsListFilterParam = `payments-list-filter-param`;
 export const paymentIdPaymentsListFilterParam = `${paymentsListFilterParam} payment-id-payments-list-filter-param`;
 export const terminalIdPaymentsListFilterParam = `${paymentsListFilterParam} terminal-id-payments-list-filter-param`;
 export const paymentStatusPaymentsListFilterParam = `${paymentsListFilterParam} payment-status-payments-list-filter-param`;
+export const paymentModePaymentsListFilterParam = `${paymentsListFilterParam} payment-mode-payments-list-filter-param`;
+export const checkoutIdPaymentsListFilterParam = `${paymentsListFilterParam} checkout-id-payments-list-filter-param`;
+export const accountHolderPaymentsListFilterParam = `${paymentsListFilterParam} account-holder-payments-list-filter-param`;
 export const createdAfterPaymentsListFilterParam = `${paymentsListFilterParam} created-after-payments-list-filter-param`;
 export const createdBeforePaymentsListFilterParam = `${paymentsListFilterParam} created-before-payments-list-filter-param`;
 
@@ -142,6 +145,7 @@ export const createdBeforePaymentsListFilterParam = `${paymentsListFilterParam} 
 
 export const payoutsListFilterMenu = `${filterMenu} payouts-list-filter-menu`;
 export const payoutsListFilterParam = `payouts-list-filter-param`;
+export const statusPayoutsListFilterParam = `${payoutsListFilterParam} status-payouts-list-filter-param`;
 export const createdAfterPayoutsListFilterParam = `${payoutsListFilterParam} created-after-payouts-list-filter-param`;
 export const createdBeforePayoutsListFilterParam = `${payoutsListFilterParam} created-before-payouts-list-filter-param`;
 
@@ -160,6 +164,8 @@ export const terminalsListFilterParam = `terminals-list-filter-param`;
 export const terminalIdTerminalsListFilterParam = `${terminalsListFilterParam} terminal-id-terminals-list-filter-param`;
 export const terminalOrderIdTerminalsListFilterParam = `${terminalsListFilterParam} terminal-order-id-terminals-list-filter-param`;
 export const terminalStatusTerminalsListFilterParam = `${terminalsListFilterParam} terminal-status-terminals-list-filter-param`;
+export const createdAfterTerminalsListFilterParam = `${terminalsListFilterParam} created-after-terminals-list-filter-param`;
+export const createdBeforeTerminalsListFilterParam = `${terminalsListFilterParam} created-before-terminals-list-filter-param`;
 
 // Terminal Orders List Filters
 export const terminalOrdersListFilterMenu = `${filterMenu} terminal-orders-list-filter-menu`;